+++ src/App.tsx
@@ -2,21 +2,18 @@ import { BrowserRouter as Router } from "react-router-dom"
 import Header from "./widgets/ui/Header.tsx"
 import Footer from "./widgets/ui/Footer.tsx"
 import PostsManagerPage from "./pages/PostsManagerPage.tsx"
-import { PostsProvider } from "./shared/context/PostsContext.tsx"
 
 const App = () => {
   return (
-    <PostsProvider>
-      <Router>
-        <div className="flex flex-col min-h-screen">
-          <Header />
-          <main className="flex-grow container mx-auto px-4 py-8">
-            <PostsManagerPage />
-          </main>
-          <Footer />
-        </div>
-      </Router>
-    </PostsProvider>
+    <Router>
+      <div className="flex flex-col min-h-screen">
+        <Header />
+        <main className="flex-grow container mx-auto px-4 py-8">
+          <PostsManagerPage />
+        </main>
+        <Footer />
+      </div>
+    </Router>
   )
 }
